refactor: Header 컴포넌트 organisms 디렉터리로 이동
- Header 컴포넌트에 position: fixed 스타일을 Fixed 컴포넌트로 대체
