--- src/components/Header/Header.type.ts
@@ -1,5 +1,0 @@
-import { AvailableColor } from 'styles/type';
-
-export type HeaderProps = {
-  backgroundColor?: AvailableColor;
-};
