+++ src/components/organisms/Header/Header.type.ts
@@ -1,0 +1,10 @@
+import { SerializedStyles, Theme } from '@emotion/react';
+import { AvailableColor } from 'styles/type';
+
+type HeaderProps = {
+  backgroundColor?: AvailableColor;
+};
+
+type GetFixedCustomStyle = (theme: Theme) => SerializedStyles;
+
+export { HeaderProps, GetFixedCustomStyle };
